+++ packages/orchestrator/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Versioning
+
+- Upgraded `@micro-lc/composer`
+
 ## [2.4.2] - 2024-06-13
 
 ### Fixed